Renamed and moved CompositeTask serializer alongside CompositeTask class.
