--- modules/runtime/src/org/lucidj/runtime/package-info.java
@@ -1,6 +1,0 @@
-@Version ("1.0.0")
-package org.lucidj.runtime;
-
-import aQute.bnd.annotation.Version;
-
-// EOF
